Split non-core logic into src/static-layout
